Fix : Added Componets For Dashboard by Codenuclei
